Reform excerpt handling
Introduce a new kind of excerpt, introduced by an additional `+++` in
a post body.

New excerpts are also saved in Frontmatter::excerpt, but they are also
included in the post body. This allows avoiding duplication of excerpt
text in post body, when the desired excerpt is also at the beginning of
the body.

Old style-excerpts, set via frontmatter yaml, are retained. Old style
excerpts allow setting excerpt text to something that isn't the opening
text of the post.